+++ src/storage/mod.rs
@@ -119,14 +119,6 @@
 //!     ) -> Result<()> {
 //!         unimplemented!()
 //!     }
-//!
-//!     async fn md5<P: AsRef<Path> + Send + Debug>(
-//!         &self,
-//!         user: &Option<DefaultUser>,
-//!         path: P,
-//!     ) -> Result<String> {
-//!         unimplemented!()
-//!     }
 //! }
 //!
 //! impl Metadata for Meta {