+++ crates/unftp-sbe-fs/src/tests.rs
@@ -102,13 +102,7 @@ fn fs_get() {
             return Err(());
         }
         assert_eq!(data.as_ref(), &*my_content);
-        // We need a `Err` branch because otherwise the compiler can't infer the `E` type,
-        // and I'm not sure where/how to annotate it.
-        if true {
-            Ok(())
-        } else {
-            Err(())
-        }
+        Ok(())
     })
     .unwrap();
 }
@@ -234,4 +228,26 @@ fn fs_rename_dir() {
 
     let old_full_path = root.join(old_dir);
     std::fs::symlink_metadata(old_full_path).expect_err("Old directory should not exists anymore");
+}
+
+#[test]
+fn fs_md5() {
+    let root = std::env::temp_dir();
+    const DATA: &str = "Some known content.";
+
+    // Create a temp file with known content
+    let file = tempfile::NamedTempFile::new_in(&root).unwrap();
+    let filename = file.path().file_name().unwrap();
+    let mut file = file.as_file();
+
+    // Create a filesystem StorageBackend with the directory containing our temp file as root
+    let fs = Filesystem::new(&root);
+    file.write(DATA.as_bytes()).unwrap();
+
+    // Since the filesystem backend is based on futures, we need a runtime to run it
+    let rt = tokio::runtime::Builder::new_current_thread().build().unwrap();
+
+    let my_md5 = rt.block_on(fs.md5(&Some(DefaultUser {}), filename)).unwrap();
+
+    assert_eq!("ced0b2edc3ec36e8d914320cb0268359", my_md5);
 }