+++ src/server/ftpserver.rs
@@ -5,7 +5,7 @@ use super::{
     chancomms::{ControlChanMsg, ProxyLoopMsg, ProxyLoopReceiver, ProxyLoopSender},
     controlchan,
     datachan::spawn_processing,
-    ftpserver::{error::ServerError, options::FtpsRequired},
+    ftpserver::{error::ServerError, options::FtpsRequired, options::SiteMd5},
     tls::FtpsConfig,
 };
 use crate::{
@@ -70,6 +70,7 @@ where
     proxy_protocol_mode: ProxyMode,
     proxy_protocol_switchboard: Option<ProxyProtocolSwitchboard<Storage, User>>,
     logger: slog::Logger,
+    sitemd5: SiteMd5,
 }
 
 impl<Storage, User> Debug for Server<Storage, User>
@@ -140,6 +141,7 @@ where
             ftps_tls_flags: TlsFlags::default(),
             ftps_client_auth: FtpsClientAuth::default(),
             ftps_trust_store: options::DEFAULT_FTPS_TRUST_STORE.into(),
+            sitemd5: SiteMd5::default(),
         }
     }
 
@@ -608,6 +610,29 @@ where
             }
         }
     }
+
+    /// Enables SITE MD5
+    ///
+    /// Warning:
+    /// Depending on the storage backend, SITE MD5 may use relatively much memory and generate high CPU usage.
+    /// This opens a Denial of Service vulnerability that could be exploited by malicious users, by means of flooding the server with SITE MD5 commands.
+    /// As such this feature is probably best user configured and at least disabled for anonymous users by default.
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// use libunftp::Server;
+    /// use libunftp::options::SiteMd5;
+    /// use unftp_sbe_fs::ServerExt;
+    ///
+    /// // Use it in a builder-like pattern:
+    /// let mut server = Server::with_fs("/tmp").sitemd5(SiteMd5::None);
+    /// ```
+
+    pub fn sitemd5<M: Into<SiteMd5>>(mut self, sitemd5_option: M) -> Self {
+        self.sitemd5 = sitemd5_option.into();
+        self
+    }
 }
 
 impl<Storage, User> From<&Server<Storage, User>> for controlchan::LoopConfig<Storage, User>
@@ -629,6 +654,7 @@ where
             logger: server.logger.new(slog::o!()),
             ftps_required_control_chan: server.ftps_required_control_chan,
             ftps_required_data_chan: server.ftps_required_data_chan,
+            sitemd5: server.sitemd5,
         }
     }
 }