+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Upcoming
+
+_tag: TAG_
+
+- Added SITE MD5 support for Filesystem and GCS storage backends. The feature is disabled by default.
+
 ## 2021-05-02 libunftp v0.17.3
 
 _tag: libunftp-0.17.3_